resume-analyzer/requirements.txt
streamlit==1.32.2
tensorflow==2.15.0
tensorflow-hub==0.14.0
pdfminer.six==20221105
spacy==3.6.1
scikit-learn==1.3.1
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
